Move scaling http codes from core to bootstrap.
